feat: 【个人】配置默认词典，抽取收藏单词组件

+++ client/src/helper/Storage.ts
@@ -4,12 +4,18 @@ export default class Storage {
   name: string
   public data: any
 
-  constructor(name: string, onSave?) {
+  constructor(name, {readOnly, onSave}: {readOnly?: boolean, onSave?: Function} = {}) {
     this.name = prefix + name
     const json = localStorage.getItem(this.name)
-    this.data = json ? JSON.parse(json) : {}
+    this.data = (json && json !== 'undefined') ? JSON.parse(json) : {}
+    if (readOnly) {
+      return
+    }
     window.addEventListener('beforeunload', () => {
-      const data = onSave?.(this.data)
+      let data = this.data
+      if (onSave) {
+        data = onSave(data)
+      }
       localStorage.setItem(this.name, JSON.stringify(data))
     })
   }
